transaction_verification/requirement.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http